refactor(bridge): extract message-action census formatting
